+++ upgrade/upgrade.php
@@ -1433,8 +1433,8 @@ $mysqlMainDb = ' . quote($mysqlMainDb) . ';
                             `uid` MEDIUMINT(8) UNSIGNED NOT NULL DEFAULT '0',
                             `record_start_date` DATETIME NOT NULL DEFAULT '0000-00-00 00:00:00',
                             `record_end_date` DATETIME DEFAULT NULL,
-                            `total_score` INT(11) NOT NULL DEFAULT '0',
-                            `total_weighting` INT(11) DEFAULT '0',
+                            `total_score` FLOAT(5,2) NOT NULL DEFAULT '0',
+                            `total_weighting` FLOAT(5,2) DEFAULT '0',
                             `attempt` INT(11) NOT NULL DEFAULT '0',
                             `attempt_status` TINYINT(4) NOT NULL DEFAULT '1',                            
                             `secs_remaining` INT(11) NOT NULL DEFAULT '0')
